# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rheo API",
    "description": "Rheo Sovereign AI Voice Studio & Neural Speech Synthesis Engine \u2014 Architected by masterz1311",
    "version": "0.5.0"
  },
  "paths": {
...
    "/channels": {
      "get": {
        "summary": "List Channels",
        "description": "List all audio channels.",
        "operationId": "list_channels_channels_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AudioChannelResponse"
                  },
                  "type": "array",
                  "title": "Response List Channels Channels Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Channel",
        "description": "Create a new audio channel.",
        "operationId": "create_channel_channels_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioChannelCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioChannelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/channels/{channel_id}": {
      "get": {
        "summary": "Get Channel",
        "description": "Get an audio channel by ID.",
        "operationId": "get_channel_channels__channel_id__get",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Channel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioChannelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Channel",
        "description": "Update an audio channel.",
        "operationId": "update_channel_channels__channel_id__put",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Channel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioChannelUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioChannelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Channel",
        "description": "Delete an audio channel.",
        "operationId": "delete_channel_channels__channel_id__delete",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Channel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/channels/{channel_id}/voices": {
      "get": {
        "summary": "Get Channel Voices",
        "description": "Get list of profile IDs assigned to a channel.",
        "operationId": "get_channel_voices_channels__channel_id__voices_get",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Channel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Set Channel Voices",
        "description": "Set which voices are assigned to a channel.",
        "operationId": "set_channel_voices_channels__channel_id__voices_put",
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Channel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelVoiceAssignment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioChannelCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "device_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Device Ids"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "AudioChannelCreate",
        "description": "Request model for creating an audio channel."
      },
      "AudioChannelResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "device_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Device Ids"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "is_default",
          "device_ids",
          "created_at"
        ],
        "title": "AudioChannelResponse",
        "description": "Response model for audio channel."
      },
      "AudioChannelUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "device_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Ids"
          }
        },
        "type": "object",
        "title": "AudioChannelUpdate",
        "description": "Request model for updating an audio channel."
      },
...
      "ChannelVoiceAssignment": {
        "properties": {
          "profile_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Profile Ids"
          }
        },
        "type": "object",
        "required": [
          "profile_ids"
        ],
        "title": "ChannelVoiceAssignment",
        "description": "Request model for assigning voices to a channel."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}